Removed custom colors for the Variations panel
There was some custom code to change the background color of the cells
if the value was different than zero.

However, for some reason I don't understand (and I don't want to
investigate deeper), this logic doesn't work well, and the background
seems the same regardless of the value.

To make it worse, since only the background is changed, the text becomes
unreadable if the user is using a dark theme for the desktop
environment. See #45.

This commit removes this custom logic. I don't think it's bringing any
value (since it's not working for me, the color is always the same) and
it is causing issues. The removal of this logic makes this panel follow
the user's theme, which makes the application look nicer and more
cohesive.

Fixes #45.