Fix unreadable variations widget text with dark themes
Constant colors values used to initialize QBrush instances returned as a
background ItemDataRoles from the VarsTableModel results in a text color
that closely matches the background color with some themes.

Return QBrush instances for Base, Text, Highlight, and HighlightedText
ColorRoles from the QApplication::palette as VarsTableModel data for
background and foreground roles to match the active theme.

This problem was reported in issue #45.

https://github.com/bitsed/qosmic/issues/45