Fix unreadable variations widget text with dark themes
Constant colors values used to initialize QBrush instances returned as
background ItemDataRoles from the VarsTableModel result in a text color
that closely matches the background color with some themes.  This
problem was reported in issue #45 with a proposed solution in #46.

Use the QApplication default bold font and Base color to render non-zero
variation value text, and use the AlternateBase color for the
background.

Also, horizontally align the "x" reset text in column 2.