Use QLocale() to format chaos and variations widget value
Values managed by Variations and Chaos widgets do not work with locales
that use a comma separator in floating point representation.  The
program is not correctly or efficiently managing this information in the
data models.

The solution is to use the application QLocale() to convert between a
character representation for floating point values and binary when
accepting user input or displaying output visible to the user.  This
solves issue #43.

Additional changes,
 Represent flam3 genome memory using references stored in the
 ChaosTableView's data model.  This avoids copying of values into the
 data model while editing.

 Remove ChaosWidget::chaosEditedSlot, and connect the
 ChaosTableView::valueUpdated signal directly to the MainWindow::render
 slot via the ChaosWidget::dataChanged signal.

 Add a function ChaosTableView::resetChaosValue that is called by event
 handlers for Delete, Backspace keys, and MiddleClick mousePressEvent.
 The resetChaosValue sets a chaos value to default of 1.0.

 Add ChaosTableView::keyPressEvent event handlers for keys 0 through 9,
 Comma, Period, Plus, Minus, and Space that opens the QTableModel editor
 for the current index.  This allows quick editing of values using the
 keyboard.  Also add Up and Down key handlers with key modifiers
 that increase or decrease a value by 10, 1, or 1/10 decimal
 precision.