Improve VarsTableModel representation of variations
Variation and variable values have been represented as character strings
in the VarsTableModel.  Every edit of a value involved a conversion from
string to binary double, write to memory, and then convert back to a
character string.  This change eliminates one string to binary
conversion and handful of memory dereference operations by storing a
reference to the selected xform's variation and variable memory in the
VarsTableModel along with the string representation given by the
application's QLocale().

Additional changes,
 Add VarsTableWidget::keyPressEvent event handlers for keys 0 through 9,
 Comma, Period, Plus, Minus, and Space that open the QTableModel editor
 for the current index.  This allows quick editing of values using the
 keyboard.  Also add Up and Down key handlers with key modifiers
 that increase or decrease a value by 10, 1, or 1/10 decimal
 precision.

 Fix unreadable variations widget text with dark themes.  An updated
 issue #45 solution from commit 665fef0920.  The default bold font and
 AlternateBase background color are used to render a variation or
 variable label, its value, and the 'x' button in the same row when the
 value is non-zero.

 Remove '_' and other punctuation from variation and variable labels.